% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_cost_cd.R
\docType{data}
\name{housing_cost_cd}
\alias{housing_cost_cd}
\title{Canadian housing affordability by census division}
\format{
\subsection{\code{housing_cost_cd}}{

A data frame with 293 rows and 4 columns:
\describe{
\item{GEO}{Census subdivision name}
\item{DGUID}{UID of census subdivision}
\item{pop2021}{Population according to 2021 census}
\item{high_housing_cost_prop}{Proportion of residents who spend more than 30\% of their
income on housing}
...
}
}
}
\source{
\url{https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=9810025501}
}
\usage{
housing_cost_cd
}
\description{
The proportion of Canadian residents who spend more than 30\% of their
income on housing.
}
\keyword{datasets}
