% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_housing_cost_map.R
\name{show_housing_cost_map}
\alias{show_housing_cost_map}
\title{Show Canadian housing cost map}
\usage{
show_housing_cost_map()
}
\value{
a leaflet map
}
\description{
Show Canadian housing cost map
}
